Created more Endpoint Functions #40
1. Admin
2. Bed
3. Bill
4. Department
5. DietPlan
6. Doctor

+++ data-access/src/main/java/com/healthplus/dataaccess/repo/BillRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.healthplus.dataaccess.domain.Bill;
 
-public interface BillRepository extends JpaRepository<Bill, Integer> {
+public interface BillRepository extends JpaRepository<Bill, Long> {
     @Query(value="Select * From Bill Where patient=?1", nativeQuery=true)
 public 	List<Bill> getBillByPatient(Integer id);
